Fix broken import paths from refactoring
