+++ src/views/user/UserDetails.js
@@ -1,7 +1,7 @@
 import { Image } from 'react-bootstrap';
 import React from 'react';
 import fetch from 'isomorphic-unfetch';
-import config from '../shared/config';
+import config from '../../config';
 import Layout from '../shared/components/Layout';
 import Order from '../orders/components/OrderCard';
